- (for 1.6.3:) unbound.h exports the shm stats structures.  They use   type long long and no ifdefs, and ub_ before the typenames.
git-svn-id: file:///svn/unbound/trunk@4117 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ util/timehist.c
@@ -225,7 +225,7 @@ timehist_quartile(struct timehist* hist, double q)
 }
 
 void 
-timehist_export(struct timehist* hist, size_t* array, size_t sz)
+timehist_export(struct timehist* hist, long long* array, size_t sz)
 {
 	size_t i;
 	if(!hist) return;
@@ -236,7 +236,7 @@ timehist_export(struct timehist* hist, size_t* array, size_t sz)
 }
 
 void 
-timehist_import(struct timehist* hist, size_t* array, size_t sz)
+timehist_import(struct timehist* hist, long long* array, size_t sz)
 {
 	size_t i;
 	if(!hist) return;
